requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
python-dotenv
google-genai
httpx
pytest
pytest-asyncio
